work

+++ ssz-rs-derive/src/lib.rs
@@ -469,7 +469,7 @@ fn derive_merkle_proof_impl(
                 fn get_proof(&mut self, idx: Vec<u64>) -> serde_json::Map<String, serde_json::Value> {
                     let roots = self.get_hash_tree();
                     // directions, val, root_bytes, proof
-                    let mut proof = ssz_rs::__internal::get_list_proof(roots, idx[0]);
+                    let mut proof = ssz_rs::__internal::get_list_proof(roots, idx[0] as usize);
 
                     // let mut index = idx.clone();
                     // let mut index = idx[0];
